Merkle Explorations
This doesn't work well, because to slash accounts, I must traverse the whole tree, extract the array at each leaf, check if it exceeds the threshold and so on. It is not more efficient in this case than a linked list and adds a lot of complexity. We definitely want guardians to be able to submit arrays of accounts, as sybils almost always exist as rings, so if you find one, you find many. However, arrays of accounts as a single leaf (required by the fact that the proof is already an array) just isn't fun to work with.